Behaviours: add key-tap behavior
This is a solution for keys that use the same keycode but different
modifiers (see #1076 as an example, also { and [ or ] and } might occur
quite often).

The new key behavior introduced here is a simple tap that immediately
sends a release event, so that a keypress of the same keycode with a
different modifier can be detected immediately.